feat:system: add Pgroup
Add a Pgroup object to represent a group of process to wait or trywait.
This object makes possible to wait for the first process of a group to
terminate or to wait for many processes without getting stuck because
some of them are waiting for you to read from their piped stdout.