feat:fleet: change the STDxxx flags to STDxxxS for execute
Change the `STDIN`, `STDOUT` AND `STDERR` flags for `Fleet->execute()`
so they now are named `STDINS`, `STDOUTS` and `STDERRS` and have a
different semantic: if the value is an 'ARRAY', then it must be the
size of the fleet and each element is the value given to the `STDIN`
(and so on) of the members execution. If the value is not an 'ARRAY',
then it is given to the `STDIN` (and so on) of every members.