feat:fleet: change the TARGET flag to TARGETS for send
Change the `TARGET` flag for `Fleet->send()` so it is now named
`TARGETS` and has a different semantic: if the value is an ARRAY, then
it must be the size of the fleet and each element of the value is given
as the `TARGET` to the members. If the value is not an ARRAY the
semantic remains the same.