feat:fleet: add a recv method to the fleet
Add a `Fleet->recv()` method which works the same as for
`Worker->recv()` except that the `TARGET` option is replaced by a
`TARGETS` option which also accepts an ARRAY as value. In this case,
each element of the ARRAY is transmitted to the members when calling
their `Worker->recv()` methods.