fix:run: partially fix a stream capture for remote scripts
The `Minion::Run::Runner` was capturing streams of each worker execution
when running a remote command. Since all the worker executions are
wrapped inside a forked process, the capture was happening in this
process and not in the main one. Now the capture happens in the main
process.

A per-worker capture is not yet implemented. It would require to setup
some `Minion::System::Process->execute(IO => [ ... ])` magic.